+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.6.1-dev
  - replace --print-stats with --no-stats, default to printing stats
  - make gaggle an optional compile-time feature
+ - GooseState is now GooseAttack
 
 ## 0.6.0 May 14, 2020
  - Initial support for gaggles: distributed load testing